split large file, but may revert
